Search both sorts: opened and updated
Neither are perfect, and when there are two PRs:
1. A long-open one that closed recently.
2. A shorter-open one that closed longer ago.

Then sometimes the long-open one will appear to have taken a long time
to publish since its merge and publish dates are recent but it doesn't show up on the
list of recently-*opened* PRs. So we'll end up using the old-ish merge
date of the shorter-open PRs.